nexus: flesh out migration saga
Flesh out the live migration saga. See RFD 361 and the theory statements in
instance_migrate.rs for more details about the design.

The new saga will start migration, but (yet) more changes are required for
it to succeed end-to-end, since sled agent still needs to be taught to send the
correct updates when a migration ends.

Remove the portion of the migration saga that prospectively reprograms OPTE
V2P mappings using the instance's destination sled. A future commit will update
V2P mappings in response to the end-of-migration instance update that records
an instance's new home sled.