NTP zones and time synchronisation (#2628)
Every sled runs an NTP zone which is started once the underlay
network is available. An NTP zone runs in one of two roles - either as
a boundary server or non-boundary. A boundary server is responsible
for synchronising time with the external network, whereas a
non-boundary server synchronises with the boundary servers.

The basic flow of time synchronisation is, per RFD34,

```
                       +----------+
                       | External |
                       |  Source  |
                       +----------+
                         /      \
                       /          \
                     v              v
               +----------+    +----------+
               | Boundary |    | Boundary |
               |   NTP    |    |   NTP    |
               |  Zone    |    |  Zone    |
               +----------+    +----------+
                   |||             |||
                   vvv             vvv
     +--------+ +--------+    +--------+ +--------+
     | Gimlet | | Gimlet |    | Gimlet | | Gimlet |
     |   NTP  | |   NTP  |....|   NTP  | |   NTP  |
     |  Zone  | |  Zone  |    |  Zone  | |  Zone  |
     +--------+ +--------+    +--------+ +--------+
```

Several services on a sled should be held back until time is
synchronised "enough".

#### Boundary NTP zone

A boundary NTP zone needs a way to communicate with the external
source. It also needs to be able to perform DNS resolution as it is
likely that the external source will be a DNS name resolving to
multiple time servers. This traffic is all outbound, so could be
behind NAT.

#### Non-Boundary NTP zone

A non-boundary NTP zone synchronises time with the boundary NTP zones.
It discovers these via the internal DNS.