Run mg-ddm in the switch zone in transit mode
Install the mg-ddm zone image into the switch zone, configure it to run
in transit mode, and listen on either all the rear tfport links or a
delegated list of physical links (for non-gimlet based testing).

RSS now asks the switch zone mg-ddm for a list of peers before creating
a sled and service plan.

Fixes #2562.

Prereqs before merging:

    XXX there's an open PR changing Nexus to use OPTE that removes this
    XXX ddmd currently panics if a link is not ready
    XXX rearN soon