[mgs] Add /local/switch-id endpoint (#2783)
This adds a `/local/switch-id` endpoint, which will return either

```json
{"type":"switch","slot":0}
```

or

```json
{"type":"switch","slot":1}
```

identifying the physical location of the switch this MGS instance is
connected to. Additional details about the switch's SP (serial number,
etc.) can then be fetched using the existing endpoints (e.g.,
`/sp/switch/0` or `/sp/switch/1`).

Note that like all other MGS endpoints, this requires MGS's location
discovery process (RFD 250) to have completed; on "human scale" this is
usually quite fast, but if dendrite is using this early in switch zone
bringup it may be able to ask MGS before MGS knows an answer (in which
case it will get back an error response and will need to retry).