[wicket/wicketd] Add ignition control (#2776)
This PR adds a few related-but-not-entirely-the-same things; if I need
to break this up into separate bits (or only keep some of them), that's
fine. In order from "most sure it's fine to merge" to "lease sure":

1. Adds an ignition command endpoint to `wicketd`; this just acts as a
proxy for the same endpoint in MGS.
2. Adds a `force_refresh` query parameter to `wicketd`'s `get_inventory`
endpoint. If the caller passes `true`, wicketd will immediately trigger
a poll to MGS and only return a response to the caller after the next
MGS poll comes back instead of returning the latest cached data.
3. Adds some mechanics in `wicket` to use `force_refresh` after a user
interaction that we know should change the inventory state. Currently
this is only used by 4.
4. Adds an ignition control popup to `wicket`. This is reachable by
pressing `<i>` on either the inventory or update screens:


![ignition-popup](https://user-images.githubusercontent.com/1435635/230470012-719def88-b527-472a-afa5-4f436356a468.png)

I'm _very much_ not sold on the UI here, but wanted to put something in
for now. One major flaw currently is that there is no place for feedback
(either success or failure); pressing enter sends the command to
`wicketd`, but results can only be observed indirectly (e.g., by seeing
the ignition power state in the inventory screen change).