Flatten selectors; allow owned names to be looked up (#2716)
This is a resurrection of #2647 with a simplified approach as originally
suggested by @sunshowers. Instead of the approach I took last time, this
introduces `OwnedName` in the lookup macro such that you can lookup
resources via `_name` or `_name_owned`. This added flexibility enables
the simplified selector approach (which I'll discuss more below) and
allows us to pass owned names in a given scope when it may otherwise be
impossible.

I've endeavored to change as little as possible and reduce the overall
noise.

---

The selector approach that I implemented in the V1 work is nice in that
it's compositional. The niceness of it sort of ends there though. It has
an over-reliance on serde's flatten. It's not the easiest to see all the
properties that a selector contains. On top of that it's not possible to
individually document parameters that are inherited from a parent
selector. This point is important because our hierarchy is only explicit
through error handling. I think it's important to be able to communicate
through our docs that a certain parameter depends on another. Given that
each parameter is _contextual_ to the selector in which it lives it's
not possible to give unique docs to each given their implementations.

@davepacheco noted in the last PR
> ...with the nested "Russian doll" approach, the hierarchy wasn't
duplicated in each struct, which seems better? (It seems like lots of
opportunities for error to duplicate these fields in a lot of places.)
The previous approach presumably makes it easier to write functions that
operate on the selector for some middle node in the hierarchy (like
Vpc), whereas it seems like you can't do that today because things
further in the hierarchy don't use a VpcSelector any more?

That feedback is still relevant but I think it's important to keep in
mind the role that selectors play in our system. A selector is
constructed at the API layer and _should not_ be passed to the app layer
(beyond performing lookups). We specifically want lookups to be the
primitives that we're passing around at the app layer. Given their
intentional limited exposure I feel like the flattened selectors don't
pose any outsized risk in producing error prone code. The makeup of the
selector itself is enforced by its lookup (which every selector has) and
the lookups themselves are recursive (such that any selector only cares
about its principle parameter and delegates the remaining parameters to
the next lookup handler in the chain).

---

I also noticed a few minor inconsistencies in some recently added
endpoints so I fixed those as well.