[wicketd] Poll MGS's individual SP endpoint instead of the bulk endpoint (#2808)
This reworks wicketd's internal polling of MGS; it now spawns N+1 tokio
tasks (where N is the number of SPs in the rack, and the +1 is for
polling ignition state), each of which is responsible for polling MGS on
some interval (which can vary over time, based on whether ignition
reports that that SP is present) for the state of its single target SP.
The end result of this is that `wicket`'s inventory display is much
snappier, particularly when the user interacts with ignition: we now see
an update in the UI nearly instantaneously, whereas previously it could
take several seconds.

There is one small change to wicketd's API: the `ignition` and `state`
fields of the SP inventory are now both optional, as they're being
polled separately and either could be missing if the other is present.
The way it is currently implemented, it's not possible for wicket to
observe an inventory response where both ignition and state are `None`
(i.e., at least one of them is guaranteed to be `Some(_)`). This isn't
represented in the type system but I'm not sure it necessarily needs to
be? Currently wicket only dumps this data as a debug string anyway - if
we need a stricter representation in the future, we could make that
change.

This PR adds one small endpoint to MGS: returning a list of all SPs MGS
is configured to talk to. This way `wicketd` doesn't need to hardcode
"32 gimlets, 2 pscs, 2 sidecars" to know how many tasks to spawn: it can
just ask MGS.