Fix bash variable collision (#2832)
If the command `retry` calls succeeds but sets `$rc` to a value other
than 0, `retry` was seeing that overwritten value of `$rc` and assuming
the command failed. `install_packages` in particular uses `$rc` to catch
"exit code 4 means no updates are necessary":
https://github.com/oxidecomputer/omicron/blob/7d6eadfd6c5cf5c0e72796c98bd711dc35a9e1e6/tools/install_builder_prerequisites.sh#L142-L144

I'm not sure if there's a _cleaner_ fix to this (some way to scope
variables?), but just changing the name solves the immediate problem.