Use Propolis ID as VM sled resource key
Using instance IDs to reserve VMM resources on sleds is not quite flexible
enough, because a single instance can have multiple Propolis VMMs. Use the
Propolis ID as the key for these instead.

Add a function to allow the deletion saga to fetch a previously-deleted instance
record so that it can obtain this ID at the correct time. Simply returning the
deleted record from the delete-record step is insufficient because the record-
deleting step needs to be idempotent and, if it runs more than once, may not
find any record to delete and return.

Tested: cargo test. Will do some ad hoc testing before merging.