Introduce Propolis ID generations (#2836)
Live migration introduces a way for an instance's Propolis ID, sled ID,
server IP, and migration information to change on the fly. To update
this state safely, introduce to the instance runtime state the concept
of a "Propolis generation" that governs how an instance's current
"location" is allowed to change.

An instance's runtime state can update from A to B if

- A and B have the same Propolis generation and Propolis ID, and B has a
later runtime state generation than A, or
- B has a later Propolis generation than A.

This creates an orderly way to change an instance's "Propolis pointer"
during a migration without forcing the newer Propolis's sled to worry
about the state generations the old sled was supplying. See RFD 361 for
more details.

This change just defines the new column. It will come into regular use
in future changes.