sim sled agent: allow direct access to simulated instances (#2861)
Wrap `SimInstance`'s current state in an `Arc`/`Mutex` so that
`SimInstance`s can be `Clone`. Then add a `SimCollection` impl that
returns a clone of one of the objects in the collection.

This allows simulated sled agent code sitting above `SimCollection` to
access a `SimInstance` in its collection and call methods on directly
(with side effects that are visible to other users of the collection).
Previously, the sled agent was more or less restricted to calling
generic `Simulatable` methods for which `SimCollection` provided an
appropriate proxy.

This allows the `InstanceStates` struct to implement interesting methods
(e.g. setting and clearing migration IDs) that regular and simulated
sled agents can share (previously such methods had to map to generic
functionality in the `Simulatable` trait). Subsequent changes will use
this heavily. For now, use it to clean up a bit of other simulated
instance code.