reserve resources on migration destination sleds
Add the ability to constrain sled selection to a specific set of desired target
sleds. This uses a builder that can be extended in the future for other
purposes (e.g. switching to "don't schedule to this sled," "schedule only to
sleds having such-and-such property," and so forth). Use this to add a
resource-reservation step to the migration saga.

Also fix a simulated sled agent bug that made simulated unregistration not
idempotent.