[update-engine] add the ability to create nested engines (#2829)
For the installinator, we need to be able to break a step into a series
of sub-steps. That is because while writing to disk, we need to handle
multiple different kinds of progress events:

* once for each M2 slot
* once for each artifact
* with retries

To handle these cases, allow for an individual step to create update
engines of its own. Parent steps can create as many child engines as
necessary, with as many steps as necessary.

This is not quite fully dynamic/monadic -- it would be so if we
flattened all events at the top level, but we maintain a hierarchy which
is really nice in practice, since it allows recording elapsed times in a
hierarchical fashion.

To handle nested engines, we also need to introduce the ability to
represent a generic spec that can represent any arbitrary value. I've
added that via a new GenericSpec spec and a NestedSpec type alias.
NestedError specifically represents a set of serialized errors sent over
the wire -- I haven't done the work to plug this in so the exact details
might change. (Most of this PR is boilerplate to support generic specs.)

I've managed to use nested engines in the installinator (PR #2830),
and they work super well! I've also updated the example to
use nested engines, though I didn't have the time to update the UI to
display nested events. It is totally possible to do this, just a bit of
additional work that I want to come back to in a future PR.