Implement silo level images (#2772)
Implements silo level images as a replacement for global images. 

This implementation preserves a single API endpoint `/v1/images` and a
single database table to store all images. It uses database views like
#2767 to map the notion of `ProjectImage` and `SiloImage` to the images
table. From an authz perspective there are actually _three_ resources
represented:`SiloImage`, `ProjectImage`, and `Image`. As the names
suggest, `SiloImage` is the child of a silo and `ProjectImage` is the
child of a project. `Image` occupies an odd space where it's currently
considered the child of a Silo though it technically straddles both
positions in the hierarchy.

---------

Co-authored-by: David Crespo <david-crespo@users.noreply.github.com>
Co-authored-by: David Crespo <david@oxidecomputer.com>