Set failmode=continue on zpools in sled-agent.
see https://github.com/oxidecomputer/omicron/issus/2766 for details.
The short version is, each pool is only backed by one vdev. There is
no recovery if one starts breaking, so if connectivity to one dies
it's actively harmful to try to wait for it to come back; we'll be
waiting forever and get stuck. We'd rather get the errors so we can
deal with them ourselves.