Do not create a link local for tfpkt links
Change `links_needed` to return whether or not a link needs a link local
address in the zone. This was required for non-gimlet testing but broke
gimlets - see: oxidecomputer/stlouis#391.

This also changes `links_needed` to check all service requests for
required links, not just stopping at the first one found.