nexus: revamp live migration saga (#2892)
Revamp the live migration saga to implement (more or less, anyway) the
design described in RFD 361. See that RFD and the theory statements in
`instance_migrate.rs` for more details.

The new saga starts migrations, but more sled agent and Nexus changes
are needed for migration to succeed end-to-end. Specifically, sled
agent needs to be taught to send the correct updates when a migration
ends, and Nexus needs to be taught to handle those updates; see below
for more commentary.

Remove OPTE V2P mapping management from the migration saga proper.
This will come back in a separate change that detects when an instance
has moved between sleds and updates the mappings. This will fix the
"can't roll back V2P mappings if the migration ultimately fails" issue
identified in the old code's comments.

To maintain the shape of the existing external live migration API
(which accepts an explicit destination sled), extend the sled resource
reservation logic to accept a "must select from the following sleds"
constraint. (This is not meant to be a final decision about the shape
of the migration API; this is just meant to keep from piling still
more changes into an already large PR.)

Also fix a couple of small bugs in prior PRs:

- Make simulated sled agent's instance unregistration routine
  idempotent.
- If a call to set/clear an instance's migration IDs fails due to a
  generation number conflict, return 409 Conflict when the error is
  detected, then make sure this is properly returned to Nexus.
  (Previously, this would return 400 Bad Request at the detection
  point, but sled agent would convert that error to 500 Internal
  Server Error as part of its final internal-error-to-external-error
  From impl.)