[wicketd] Refuse to update the sled where wicketd is running (#2940)
wicketd now refuses to update the sled on which it is running, by
comparing the baseboard info it retries from the SP of the to-be-updated
sled against the baseboard info of its own host. To do this, we plumb that
baseboard info through sled-agent down into wicketd. A nontrivial
byproduct of this is the addition of `#[serde(skip)]` on several
`ServiceType` variants that should not be passed in from external
clients (i.e., services that sled agent itself is responsible for
starting autonomously, either in the gz (maghemite) or in the switch
zone (the rest)).