nexus: don't error on omitted signing_keypair SamlIdentityProviderCreate (#2947)
A different behind the scenes demo fix. The console doesn't have fields
for providing a signing keypair but it's meant to be optional.

The default deserialization behaviour is to set a field to it's default
value if it's not present so we'd expect `None` if the request didn't
include `signing_keypair`. But `#[serde(deserialize_with = ...)]`
changes that subtly such that `None` is present but empty / null. We can
use `#[serde(default)]` to make it more permissive. I don't think we
need to distinguish between "not present" vs "present but empty/null".