Support upserting multiple non-gimlet sled agents
Sled baseboard is used as a unique identifier for sleds, but for i86pc
this was set to ("unknown", "unknown", 0), and this caused sled upsert
to fail due to "object already exists".

Change Baseboard from a struct to an enum, preserving the type of the
baseboard along with its identifying information. Serialize this to a
file for wicketd to consume instead of supplying it via command line
arguments.

Generate a unique serial number for i86pc baseboards during hardware
view snapshot.