[tufaceous-lib] More general support for composite artifacts
We previously had some support for two kinds of composite artifacts:

* Host/Trampoline OS images (phase1/phase2)
* RoT archives (A/B images)

but it had grown a bit organically. This PR attempts to add more general
support for defining composite artifacts in the manifest used by
`tufaceous assemble`, and adds support for the control plane composite
artifacts (which holds an arbitrary number of control plane zone
tarballs).