[wicket] correctly wrap popup text within step logs (#3081)
Previously, we were relying on tui's paragraph wrap functionality. This
had several limitations:

1. There's no way to obtain the actual height of the text as computed.
   This meant that wrapped text was previously cut off.
3. There's no sensible way to implement scrolling.

To address these issues, we implement wrapping ourselves, via the
textwrap library. The library exposes support for wrapping strings, but
we're working with tui Spans (which have styles associated with them) so
we have to do some work ourselves.  Thankfully, it isn't too much work:
textwrap exposes an abstract interface called `Fragment`, so it's mostly
a matter of copy-pasting some code from textwrap, and annotating each
word with its corresponding tui style.

This PR does not implement scrolling, but now that we're working with
pre-wrapped text in the popup, scrolling should be pretty simple to do
in a followup.

Screenshot of wrapped text:


![image](https://github.com/oxidecomputer/omicron/assets/180618/fb204b3a-3917-4e9d-bec8-4edb88c6cfdd)