nexus: update instance networking config after live migration (#3127)
Whenever Nexus gets a new instance runtime state from a sled agent,
compare the state to the existing runtime state to see if applying the
new state will update the instance's Propolis generation. If it will,
use the sled ID in the new record to create updated OPTE V2P mappings
and Dendrite NAT entries for the instance.

Retry with backoff when sled agent fails to publish a state update to
Nexus. This was required for correctness anyway (see #2727) but is
especially important now that there are many more ways for Nexus to fail
to apply a state update. See the comments in the new code for more
details.

In the future, it might be better to update this configuration using a
reliable persistent workflow that's triggered by Propolis location
changes. This approach will require at least some additional work in
OPTE to assign generation numbers to V2P mappings (Dendrite might have a
similar problem but I'm not as familiar with the tables Nexus is trying
to maintain in this change).