Remove "from_start" parameter from serial stream.
Removal of an API footgun, as this parameter really only makes
sense for the history-retrieval GET endpoint, not the continuous
websocket connection.

Eventually over an instance's lifespan, it makes less sense for a
client connecting to try asking for bytes starting from the
beginning of time, as we don't buffer everything forever and this
eventually becomes impossible to satisfy.