Spawn a task to run services_ensure (#3140)
When a client request times out or is otherwise aborted, the future
corresponding to a dropshot endpoint's task is dropped, which leads to
the plug being pulled on whatever is currently executing. In the case of
`services_ensure`, zones were left in a partially configured state, and
the bookkeeping in ServiceManager was incorrect. When the next call to
`PUT /services` is made, `services_ensure` will eventually try to bring
up the same zone it was interrupted at, leading to configuration issues.

Fix this by spawning a separate task to run `services_ensure`.

Fixes #3098