[wicket] show errors that occur while starting an update (#3143)
This took a couple tries to do but it works quite well now.

This PR bundles up a few things:

1. **Support for showing errors that occur while starting an update.**
We now track this state in the popup.

Note that state for errors that occur while starting an update isn't
stored in wicketd, just in wicket. I originally went down the path of
trying to store that state in wicketd, but I realized that errors while
starting an update can also occur during wicket -> wicketd communication
so gave up on that idea. (We can revisit this in the future if we need
to.)

The error messages suck at the moment (raw HTTP errors) but they're
communicative enough. I'd love for us to return more structured error
messages in the future, similar to the errors returned by the update
engine.

(See PR for screenshots.)

2. **A few changes to how start update options work.** We now upload the
options in the body of the POST request, and also add new options to
simulate failures and timeouts while starting an update. (I later
realized I could test those cases a different way, but I'd already done
the work by then and the refactoring seems useful enough.)

3. **Show a help bar at the bottom on some update screens.** We show the
help bar on:
    - The "start update" screen.
    - But not on the "waiting for events" or (currently) the event
display screen.

4. **Some refactoring of documentation.** Testing out wicket as a
captive SSH shell is a pretty rare occurrence, so de-emphasize it in the
testing documentation.