[wicket] add scrolling to popups (#3218)
Now that we're displaying error messages in popups, we can have a pretty
large amount of text here. Add the ability to scroll down in a popup,
building on #3194 which wraps all popup text.

This is a bit tricky because we want to cap the offset to prevent
scrolling down once we've reached the bottom of the text, but a lot of
values are computed at layout time.  To handle this, introduce the
notion of an "actual scroll offset", which is the scroll offset as
computed after text is laid out.  Also compute the layout parameters in
the constructor for `Popup` so we can return the actual scroll offset.

Display scroll status in the button bar (I didn't figure out a better
place to show it; we could potentially show an actual scroll bar to the
right but that's a lot more work than this.)

See https://github.com/oxidecomputer/omicron/pull/3218 for examples.